HAL is a lot mor readable now and API layer a lot more easy to understand
